Change versioning scheme
Set the version to 0.0.0.7 from 0.7.  This will allow more room for
growth.

From here, all further (significant) development will be done in
separate branches.  In those, the last digit will be incremented
on every commit that adds a feature or fixes a bug.

The third digit will be incremented on every merge from a dev
branch.

The second digit will signify a push to the chrome web store, as
well as changes that are small, but could break something.

The first digit will signify massive refactoring, as well as large
sweeping changes.