Fix input race condition
The extension was having issues in which one could type an
incorrect string match and have the search continue to match the
most recent string matched.  It turns out that the move to a
keydown event caused the search to sometimes execute before the
keystroke reached the input box.  When this happened, the search
would only search for the string, sans the new input.

This commit fixes the bug by moving the backspace and normal
typing handlers to the keyup handler, which always occurs after
the input is in the box.